5.16. Функции возвращаются PR to fork main
