+++ index.html
@@ -8,7 +8,7 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
-  <!-- <script type="text/javascript" src="js/functions.js" defer></script> -->
+  <script type="text/javascript" src="js/functions.js" defer></script>
   <script type="module" src="js/main.js" defer></script>
 </head>
