+++ js/functions.js
@@ -1,42 +1,13 @@
-const uploadFile = document.querySelector('.img-upload__input');
-const uploadOverlay = document.querySelector('.img-upload__overlay');
-const overlayCloseButton = document.querySelector('.img-upload__cancel');
-const bodyArray = document.getElementsByTagName('body');
-const [body] = bodyArray;
+// Напишите функцию, которая принимает время начала и конца рабочего дня, а также время старта и продолжительность встречи в минутах и возвращает true, если встреча не выходит за рамки рабочего дня, и false, если выходит.
 
-uploadFile.onclick = function (evt) {
-  evt.preventDefault();
-  uploadOverlay.classList.remove('hidden');
-  body.classList.add('modal-open');
-  // document.addEventListener('keydown', windowCloseEvent);
-  document.addEventListener('keydown', (e)=> {
-    if (e.keyCode === 27) {
-      uploadOverlay.classList.add('hidden');
-      body.classList.remove('modal-open');
-    }
-    // body.style.backgroundColor = 'white';
-  });
-};
-overlayCloseButton.onclick = function () {
-  uploadOverlay.classList.add('hidden');
-  body.classList.remove('modal-open');
+// Время указывается в виде строки в формате часы:минуты. Для указания часов и минут могут использоваться как две цифры, так и одна. Например, 8 часов 5 минут могут быть указаны по-разному: 08:05, 8:5, 08:5 или 8:05.
+
+// Продолжительность задаётся числом. Гарантируется, что и рабочий день, и встреча укладываются в одни календарные сутки.
+
+
+const timeToJob = (dayStart, dayEnd, meetingStart, meetingTimeMin) => {
+
+
 };
 
-
-// };
-/*
-
-1. Загрузка нового изображения на сайт и заполнение информации о нём
-1.1. Загрузка нового изображения:
-
-выбор файла с изображением для загрузки;
-изменение масштаба изображения;
-применение одного из заранее заготовленных эффектов;
-выбор глубины эффекта с помощью ползунка;
-добавление текстового комментария;
-добавление хэштегов.
-1.2. Выбор изображения для загрузки осуществляется с помощью стандартного контрола загрузки файла .img-upload__input, который стилизован под букву «О» в логотипе. После выбора изображения (изменения значения поля .img-upload__input), показывается форма редактирования изображения. У элемента .img-upload__overlay удаляется класс hidden, а body задаётся класс modal-open.
-
-После выбора изображения пользователем с помощью стандартного контрола загрузки файла .img-upload__input, нужно подставить его в форму редактирования вместо тестового изображения в блок предварительного просмотра и в превью эффектов.
-
-1.3 Закрытие формы редактирования изображения производится либо нажатием на кнопку .img-upload__cancel, либо нажатием клавиши Esc. Элементу .img-upload__overlay возвращается класс hidden. У элемента body удаляется класс modal-open.*/
+console.log(timeToJob());
