+++ js/main.js
@@ -1,2 +1,2 @@
 import './data.js';
-import './functions.js';
+import './upload-file.js';
